Revert "Add the ability to use up and down keys, better errors with red color…"
