+++ README.md
@@ -158,24 +158,11 @@ pieces search query --mode fts
 ```
 
 Examines all words in a document to find matches to search criteria
-##### change the llm model you are using:
 
-Lists all the avaiable models on your machine with their indices to choose from 
-
-```bash
-pieces change_model 
-```
-##### change the llm model you are using:
-
-Change the model in the ask command
-
-```bash
-pieces change_model [MODEL_INDEX]
-```
 ##### Ask a question to a model:
 ** Requires quotes around question **
 
-Ask the copoilt a question it uses chatGPT3 as a defualt model to ask a question you can change the model using the change model command
+This currently only supports GPT 3.5 and it does not have working memory. Only coding questions are currently supported. To use the model's code you can copy it from the console and use the create command to create an asset using the copied code. 
 
 ```bash
 pieces ask "your question"
