--- src/pieces/commands/__init__.py
@@ -1,37 +1,0 @@
-from .cli_loop import loop, find_most_similar_command
-from .autocommit import git_commit
-from .commands_functions import (version,
-                                 save_asset,
-                                 edit_asset,
-                                 set_parser,
-                                 list_assets,
-                                 open_asset,
-                                 create_asset,
-                                 ask,
-                                 search,
-                                 delete_asset,
-                                 check_api,
-                                 set_pieces_os_version,
-                                 change_model,
-                                 run_in_loop)
-
-
-__all__ = ['loop', 
-           'find_most_similar_command',
-           'git_commit',
-           'set_parser',
-           'version',
-           'save_asset',
-           'edit_asset',
-           'list_assets',
-           'open_asset',
-           'create_asset',
-           'ask',
-           'search',
-           'delete_asset',
-           'check_api',
-           'set_pieces_os_version',
-           'set_application',
-           'change_model',
-           'run_in_loop']
-
