--- src/pieces/commands/extensions.json
@@ -1,272 +1,0 @@
-{
-    "csx": ".csx",
-    "cs": ".cs",
-    "html": ".html",
-    "htm": ".htm",
-    "shtml": ".shtml",
-    "xhtml": ".xhtml",
-    "hs": ".hs",
-    "hs-boot": ".hs-boot",
-    "hsig": ".hsig",
-    "cpp": ".cpp",
-    "cc": ".cc",
-    "cp": ".cpp",
-    "cxx": ".cpp",
-    "c": ".c",
-    "h": ".h",
-    "hh": ".h",
-    "hpp": ".hpp",
-    "hxx": ".hxx",
-    "inl": ".inl",
-    "ipp": ".ipp",
-    "ixx": ".ixx",
-    "cppm": ".cppm",
-    "csv": ".csv",
-    "doc": ".doc",
-    "docm": ".docm",
-    "docx": ".docx",
-    "exe": ".exe",
-    "gif": ".gif",
-    "ico": ".ico",
-    "jpe": ".jpg",
-    "jpeg": ".jpg",
-    "jpg": ".jpg",
-    "jpgm": ".jpgm",
-    "jpgv": ".jpgv",
-    "log": ".log",
-    "mp2": ".mp2",
-    "mp2a": ".mp2a",
-    "mp3": ".mp3",
-    "mp4": ".mp4",
-    "mp4a": ".mp4",
-    "oga": ".ogg",
-    "ogg": ".ogg",
-    "ogv": ".ogv",
-    "ogx": ".ogx",
-    "ppt": ".ppt",
-    "pptx": ".pptx",
-    "pptm": ".pptm",
-    "qt": ".mov",
-    "text": ".txt",
-    "tif": ".tif",
-    "tiff": ".tiff",
-    "txt": ".txt",
-    "wav": ".wav",
-    "weba": ".weba",
-    "webm": ".webm",
-    "webp": ".webp",
-    "xla": ".xls",
-    "xlam": ".xlam",
-    "xlc": ".xlc",
-    "xlm": ".xlm",
-    "xls": ".xls",
-    "xlsb": ".xlsb",
-    "xlsm": ".xlsm",
-    "xlsx": ".xlsx",
-    "xlt": ".xlt",
-    "xltm": ".xltm",
-    "xltx": ".xltx",
-    "pdf": ".pdf",
-    "png": ".png",
-    "rpm": ".rpm",
-    "zip": ".zip",
-    "woff": ".woff",
-    "woff2": ".woff2",
-    "svgz": ".svgz",
-    "bin": ".bin",
-    "bmp": ".bmp",
-    "eot": ".eot",
-    "gz": ".gz",
-    "jar": ".jar",
-    "mpkg": ".mpkg",
-    "ai": ".ai",
-    "eps": ".eps",
-    "dmg": ".dmg",
-    "list": ".txt",
-    "rtx": ".rtx",
-    "uri": ".uri",
-    "uris": ".txt",
-    "urls": ".txt",
-    "css": ".css",
-    "dart": ".dart",
-    "java": ".java",
-    "bsh": ".bsh",
-    "js": ".js",
-    "jsx": ".jsx",
-    "mjs": ".mjs",
-    "htc": ".htc",
-    "json": ".json",
-    "ipynb": ".ipynb",
-    "gltf": ".gltf",
-    "jsonml": ".jsonml",
-    "ps": ".ps",
-    "ssml": ".ssml",
-    "wasm": ".wasm",
-    "rtf": ".rtf",
-    "cco": ".txt",
-    "pl": ".pl",
-    "pc": ".txt",
-    "pm": ".pm",
-    "pmc": ".txt",
-    "pod": ".pod",
-    "t": ".t",
-    "xml": ".xml",
-    "tld": ".txt",
-    "dtml": ".dtml",
-    "rng": ".rng",
-    "rss": ".rss",
-    "opml": ".opml",
-    "svg": ".svg",
-    "xaml": ".xaml",
-    "sublime-snippet": ".txt",
-    "tmLanguage": ".txt",
-    "tmPreferences": ".txt",
-    "tmSnippet": ".txt",
-    "tmTheme": ".txt",
-    "csproj": ".csproj",
-    "fsproj": ".fsproj",
-    "sqlproj": ".sqlproj",
-    "vbproj": ".vbproj",
-    "vcproj": ".vcproj",
-    "vcxproj": ".vcxproj",
-    "dae": ".dae",
-    "props": ".props",
-    "targets": ".targets",
-    "xsd": ".xsd",
-    "xsl": ".xsl",
-    "xslt": ".xslt",
-    "ecma": ".txt",
-    "node": ".txt",
-    "php": ".php",
-    "php3": ".php",
-    "php4": ".php",
-    "php5": ".php",
-    "php7": ".php",
-    "php8": ".php",
-    "phps": ".php",
-    "phpt": ".phpt",
-    "phtml": ".phtml",
-    "py": ".py",
-    "py3": ".py",
-    "pyw": ".pyw",
-    "pyi": ".pyi",
-    "pyx": ".pyx",
-    "pyx.in": ".pyx",
-    "pxd": ".pxd",
-    "pxd.in": ".pxd",
-    "pxi": ".pxi",
-    "pxi.in": ".pxi",
-    "rpy": ".rpy",
-    "cpy": ".cpy",
-    "gyp": ".gyp",
-    "gypi": ".gypi",
-    "vpy": ".vpy",
-    "smk": ".smk",
-    "wscript": ".wscript",
-    "bazel": ".bazel",
-    "bzl": ".bzl",
-    "pyc": ".pyc",
-    "class": ".class",
-    "p": ".p",
-    "pas": ".pas",
-    "curl": ".curl",
-    "mcurl": ".txt",
-    "go": ".go",
-    "swift": ".swift",
-    "rs": ".rs",
-    "ts": ".ts",
-    "tsx": ".tsx",
-    "rb": ".rb",
-    "rbi": ".rbi",
-    "rbx": ".rbx",
-    "rjs": ".rjs",
-    "rabl": ".rabl",
-    "rake": ".rake",
-    "capfile": ".capfile",
-    "jbuilder": ".jbuilder",
-    "gemspec": ".gemspec",
-    "podspec": ".podspec",
-    "irbrc": ".irbrc",
-    "pryrc": ".pryrc",
-    "prawn": ".prawn",
-    "thor": ".thor",
-    "Appfile": ".txt",
-    "Appraisals": ".txt",
-    "Berksfile": ".txt",
-    "Brewfile": ".txt",
-    "Cheffile": ".txt",
-    "Deliverfile": ".txt",
-    "Fastfile": ".txt",
-    "Gemfile": ".Gemfile",
-    "Guardfile": ".txt",
-    "Podfile": ".Podfile",
-    "Rakefile": ".Rakefile",
-    "Rantfile": ".txt",
-    "Scanfile": ".txt",
-    "simplecov": ".txt",
-    "Snapfile": ".txt",
-    "Thorfile": ".txt",
-    "Vagrantfile": ".Vagrantfile",
-    "scala": ".scala",
-    "sbt": ".sbt",
-    "sc": ".txt",
-    "cmd": ".cmd",
-    "bat": ".bat",
-    "coffee": ".coffee",
-    "erl": ".erl",
-    "hrl": ".hrl",
-    "escript": ".escript",
-    "lua": ".lua",
-    "md": ".md",
-    "mdown": ".mdown",
-    "mdwn": ".mdwn",
-    "markdown": ".md",
-    "markdn": ".md",
-    "matlab": ".matlab",
-    "m": ".m",
-    "ps1": ".ps1",
-    "sh": ".sh",
-    "bash": ".sh",
-    "bashrc": ".bashrc",
-    "ash": ".sh",
-    "zsh": ".zsh",
-    ".bash_aliases": ".bash_aliases",
-    ".bash_completions": ".bash_completions",
-    ".bash_functions": ".bash_functions",
-    ".bash_login": ".bash_login",
-    ".bash_logout": ".bash_logout",
-    ".bash_profile": ".bash_profile",
-    ".bash_variables": ".bash_variables",
-    ".profile": ".profile",
-    ".textmate_init": ".textmate_init",
-    ".zlogin": ".zlogin",
-    ".zlogout": ".zlogout",
-    ".zprofile": ".zprofile",
-    ".zshenv": ".zshenv",
-    ".zshrc": ".zshrc",
-    "PKGBUILD": ".PKGBUILD",
-    "ebuild": ".ebuild",
-    "eclass": ".eclass",
-    "r": ".r",
-    "sql": ".sql",
-    "ddl": ".ddl",
-    "dml": ".dml",
-    "tex": ".tex",
-    "ltx": ".ltx",
-    "sty": ".sty",
-    "cls": ".cls",
-    "unknown": ".txt",
-    "yaml": ".yaml",
-    "yml": ".yml",
-    "toml": ".toml",
-    "tml": ".txt",
-    "Cargo.lock": ".lock",
-    "Gopkg.lock": ".lock",
-    "Pipfile": ".Pipfile",
-    "poetry.lock": ".lock",
-    "uniform_resource_locator": ".txt",
-    "custom_url_scheme": ".txt",
-    "unix_file_path": ".txt",
-    "windows_file_path": ".txt",
-    "uniform_resource_identifier": ".txt"
-}
