+++ tests/web_socket_handler_test.py
@@ -1,7 +1,7 @@
 import json
 from unittest.mock import Mock, patch
 import pytest
-from src.pieces.api.pieces_websocket import WebSocketManager
+from src.pieces.api import WebSocketManager
 
 class TestWebSocketManager:
 
@@ -31,6 +31,7 @@ class TestWebSocketManager:
         })
         ws_manager.on_message(None, sample_message)
         assert ws_manager.loading == False
+        assert ws_manager.first_token_received
 
     def test_on_error(self, ws_manager):
         ws_manager.on_error(None, "Test error")
